Django2/myproject/requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.10
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-timezone-field==6.1.0
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
kombu==5.3.5
magic-filter==1.0.12
multidict==6.0.5
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
pandas==2.2.0
openpyxl==3.1.2
Werkzeug==3.0.1
pyOpenSSL==24.0.0
django-axes==6.3.0
telebot==0.0.5

